Adjust linker script to avoid having a PT_LOAD header all the way from zero
Previously, the generated binaries had a single PT_LOAD header all the
way from 0 to the end of the generated code.

When building, the domain builder maps the entire region and
zeroes/copies the data as appropriate.  This causes a failure to build
when the header overlaps with unmapped addresses such as the HVM legacy
VGA range.

Link all binaries to load and run at 1MB to avoid the legacy VGA hole in
HVM guests.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>