PV exception entry points
Infrastructure to register the virtual IDT with Xen and get execution back
into C when an exception occurs.  The existing 32 and 64bit entry points are
mostly reused, with small adjustments for PV guests.

Most of this change is importing and implementing Xen ABI bits for PV guests.

Exceptions are currently fatal.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>