runner: Rework test selection logic
The existing logic for test selection was awkward to use and somewhat
non-intuitive.  It also had an annoying habit of accidentially running the msr
utility and causing excessive logspam.

First of all, switch the all_{categories,environments} tuples to sets, which
allows for cleaner logic during selection.  Split all_categories into default
and non-default subsets, where the non-default subset is not available unless
explicitly referenced.

Introduce interpret_selection() with the purpse of interpreting a collection
of string parameter, and returning the commplete list of TestInstances covered
by the selection.

interpret_selection() is now called before action selection, and replaces the
existing ad-hoc logic in the list and run calls.  As a result, selections are
now identical for the list and run paths.

The --help documentation is also updated to reflect the new behaviour.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>