Rework the pv-iopl test using variations
The hypercall and vmassist tests are logically separate, and while the
hypercall version has been around for a very long time, vmassist was only
introduced in Xen 4.7.

To be able to sensibly test the hypercall part on older versions of Xen, they
should be treated separately, so the skip from vmassist detection doesn't
influence the results from the hypercall test.

To test them independently, introduce variations which alter the cmdline
passed to the guest.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>