Unify and shorten the I/L/GDT paths in x86_exc_decode_ec()
The split between L/GDT selector (in hex) and IDT index (in decimal) is
awkward to read when both types of errors are seen together.  Furthermore, the
vec/sel text adds unncessary length to the message.

Switch to uniformally reporting the table index in C array syntax which is
shorter to express.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>