Print test titles earlier during startup
Certain conditions, such as failing the test_needs_fep check, result in the
framework exiting before calling test_main().  Logs from such runs are
impossible to distinguish.

Require tests to provide a test_title string, and print it as part of the
startup banner.

Update make-new-test.sh and the introductory docs to match

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>