Don't automatically recover from traps
While this property is useful for swint-emulation, it is unhelpful in other
situations, as it causes unexpected traps to get swallowed silently.

Reuse the existing extable logic (with fault == fixup) to recover, and
introduce _ASM_TRAP_OK() to introduce such an extable entry.

Adjust the swint-emulation and selftest logic following this behaviour change,
which removes the test_int3_breakpoint() case entirely.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>