arm: Add initial architecture code for arm64 and arm32
Add initial support to XTF for arm64/arm32 without
modifying existing XTF architecture.
Some of the files are just dummy files waiting to be properly
implemented step by step later on.

The purpose of this change is to add the initial code and minimal
set of files to support XTF on arm without modifying the existing
XTF design. This creates a base for further implementation.

Based-on-the-work-from: Julien Grall <julien@xen.org>
Signed-off-by: Michal Orzel <michal.orzel@arm.com>