arm: Add build support for arm64/arm32
Add support for arm64/arm32 in the build
system, modify the common headers to be
aware of the new architectures and update
the documentation.

Architecture can be set using ARCH=arm64
or ARCH=arm32.

Currently only one test: tests/example
is supported allowing the startup code head.S
to be executed ending up in an infinite loop.

Signed-off-by: Michal Orzel <michal.orzel@arm.com>