arm: Add the hypercall handling interfaces
Add the hypercall wrappers for arm64 and arm32 from Linux.
The hypercalls are defined in the assembly files hypercall.S and they
are specific to arm64/arm32.

The prototypes and higher level wrappers are defined in
arch/arm/include/arch/hypercall.h as they are common to both arm64
and arm32.

Signed-off-by: Michal Orzel <michal.orzel@arm.com>