build: Modify build system to support multiple architectures
Currently XTF build system is strictly made up for x86.
Modify the build system so that it will be easier
to add other architectures to XTF in the future.
This is done by generalizing the common makefiles to be
architecture independent and creating architecture
specific makefiles.

Introduce new variable SUBARCH which is the name of
the architecture devoid of information about the bitness,
that expands as follows:
-on x86: x86
-in the future on arm64/arm32: arm

Modify the framework so that the test name contains
the SUBARCH name e.g.:
test-x86-pv64-example
instead of:
test-pv64-example

This change will require a modification in OSSTEST
due to the change in test naming.

Signed-off-by: Michal Orzel <michal.orzel@arm.com>