arm: Add support for PV console
... so that we can connect to guest console using
xl console. This change is required for OSSTEST which
makes use of xtf-runner that connects to the guest
console in order to obtain the results.

PV console should not be used when XTF is running as dom0.
For that purpose, implement a function is_initdomain
that uses hypercall xen_version to get Xen features
and checks if bit XENFEAT_dom0 is set.

Signed-off-by: Michal Orzel <michal.orzel@arm.com>