arm64: Add BSS zeroing
Add BSS zeroing into arm64 startup code head.S and implement
a function clean_and_invalidate_dcache. Image protocol states
that the state of the cache for BSS region is not known as BSS
is not part of the loaded image. This is why it is required to
invalidate the cache before and after zeroing BSS.

Signed-off-by: Michal Orzel <michal.orzel@arm.com>