runner: Use virt caps to identify whether a test can run
Right now, --host filters a test selection using xen_caps to exclude full
environments.  Furthermore, without passing --host, xtf-runner will try to
start a test even when we know it can't be constructed, leaving an xl error to
the user and reporting CRASH.

The virt capabilities are more fine grained, and include other information
such as the availability of hap and shadow paging.

Split run_test() out of run_tests() and use the test's required caps to short
circuit to SKIP.

Rework the existing --host filtering in terms of virt caps, which simplifies
the invocation of tests_from_selection().

This causes ~hap and ~shadow variation tests to SKIP rather than CRASH when
run on hardware without HAP (or with HAP disabled in firmware), or with Shadow
compiled out.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>