test-swint: Update to use ex_record rather than exlog
This is both more precise and consice.

As this test pertains to mixing up faults and traps, introduce EXINFO_TRAP and
ex_record_trap_eax() to distinguish which EXTABLE entry was used.  This in
turn means we don't need to export the trap/fault labels to C.

Change the nomeclature for instruction prefixes, because "red" is fairly
meaningless and confusing in context.

No change in test behaviour.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>